Refactor memory map workings
The way we were handling memory maps was... let's just say less than
ideal. This commit refactors out all the logic related to the memory
map (including size/start of RAM/BIOS) into `map.rs`.

Also included in `map.rs` is a convenient function that takes in an
address, does the unmasking, checks all the regions, and returns an
Enum corresponding to the matched region along with an offset.

We can now use this to allow pattern-matching on the actual region in
`bus.rs`, which makes everything so much cleaner. We also get better
debug information for free, since we can print out the unhandled regions
with the Enum!

Note: In debug mode, this is ~20% slower than the previous code, which
is sad, but the difference is much less in the optimized version. We
should profile this at some point, but I think the cleanliness is worth
it here.