Add block control registers for channels; convert some logs to trace
The RUST_LOG=debug mode was getting a bit too spammy. For now I've
moved almost all the logging to `trace` mode, since we don't actively
need them. I'm thinking here that we can set some of them to `debug`
again if we need to see certain things when debugging a particular
issue.